+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [10.1.0] 2025-05-23
+## [10.1.0] 2025-05-29
 ### Added
 - Enabled combining multiple displays and codes in a unit lookup.
+### Fixed
+- Fixed the no_match status when switching between Pick initial value and other value methods.
+- Fixed the empty answer option check for all data types except coding.
 
 ## [10.0.7] 2025-05-19
 ### Fixed